Changes till noon 31 Jul

+++ HPO/datamodule_glue.py
@@ -65,6 +65,8 @@ class GLUEDataModule(LightningDataModule):
         self.num_labels = self.glue_task_num_labels[task_name]
         self.tokenizer = AutoTokenizer.from_pretrained(self.model_name_or_path, use_fast=True)
 
+
+
     def setup(self, stage: str):
         self.dataset = datasets.load_dataset("glue", self.task_name)
 
@@ -253,11 +255,13 @@ class AmazonMultiReview(LightningDataModule):
             train_batch_size: int = 32,
             eval_batch_size: int = 32,
             label_column: str = 'label',
-            encode_columns: list = [],
+            encode_columns=None,
             **kwargs,
     ):
 
         super().__init__()
+        if encode_columns is None:
+            encode_columns = []
         self.model_name_or_path = model_name_or_path
         self.task_name = task_name
         self.max_seq_length = max_seq_length
@@ -271,12 +275,9 @@ class AmazonMultiReview(LightningDataModule):
         self.n_cpu = 2
         self.label_column = label_column
         self.encode_columns = encode_columns
+
         #  Tokenize the dataset
-        if not os.path.isfile(f'{self.task_metadata["tokenize_folder_name"]}/tokenized_data.pt'):
-            print("File not exist")
-            self.prepare_data()
-        else:
-            print("File exist. Load Tokenized data")
+        self.prepare_data_per_node = True
 
     def clean_data(self, example):
         cols = self.encode_columns
@@ -286,31 +287,33 @@ class AmazonMultiReview(LightningDataModule):
         return example
 
     def prepare_data(self):
-        # called only on 1 GPU per node. Do not do any self.x assignments here
-        # download, split tokenise data and save to disk
-        print(f'Download and Tokenise')
-        dataset = datasets.load_dataset(self.task_name, 'de')
 
-        for split in dataset.keys():
-            dataset[split] = dataset[split].map(self.clean_data, batched=True, )
-            dataset[split] = dataset[split].map(self.encode_batch, batched=True)
-            remove_features = set(dataset[split].features) ^ set([self.label_column] + ["input_ids", "attention_mask"])
-            dataset[split] = dataset[split].remove_columns(remove_features)
-            dataset[split] = dataset[split].rename_column(self.label_column, "labels")
+        if not os.path.isfile(f'{self.task_metadata["tokenize_folder_name"]}/tokenized_data.pt'):
+            print("File not exist")
+            print(f'Download and Tokenise')
+            dataset = datasets.load_dataset(self.task_name, 'german')
+            for split in dataset.keys():
+                dataset[split] = dataset[split].map(self.encode_batch, batched=True)
+                remove_features = set(dataset[split].features) ^ set(
+                    [self.label_column] + ["input_ids", "attention_mask"])
+                dataset[split] = dataset[split].remove_columns(remove_features)
+                dataset[split] = dataset[split].rename_column(self.label_column, "labels")
 
-            # Transform to pytorch tensors and only output the required columns
-            # self.dataset[split].set_format(type="torch", columns=["input_ids", "attention_mask", "labels"])
+                # Transform to pytorch tensors and only output the required columns
+                columns = [c for c in dataset[split].column_names if c in self.loader_columns]
+                dataset[split].set_format(type="torch", columns=columns)
 
-            columns = [c for c in dataset[split].column_names if c in self.loader_columns]
-            dataset[split].set_format(type="torch", columns=columns)
+            # save the tokenized data to disk
+            try:
+                Path(f'{self.task_metadata["tokenize_folder_name"]}').mkdir(parents=True, exist_ok=True)
+                torch.save(dataset, f'{self.task_metadata["tokenize_folder_name"]}/tokenized_data.pt')
+            except:
+                print("File already exist")
 
-        # save the tokenized data to disk
-        try:
-            Path(f'{self.task_metadata["tokenize_folder_name"]}').mkdir(parents=True, exist_ok=True)
-            torch.save(dataset, f'{self.task_metadata["tokenize_folder_name"]}/tokenized_data.pt')
-        except:
-            print("File already exist")
-        print(f'Dataset Tokenised')
+        else:
+            print("File exist. Load Tokenized data in setup.")
+
+
 
     def setup(self, stage: str):
         # load data here
@@ -346,6 +349,13 @@ class AmazonMultiReview(LightningDataModule):
 
 
 class TyqiangzData(LightningDataModule):
+
+    task_metadata = {
+        "num_labels": 3,
+        "label_col": "label",
+        "tokenize_folder_name": "tyqiangzdata",
+    }
+
     loader_columns = [
         "datasets_idx",
         "input_ids",
@@ -364,19 +374,19 @@ class TyqiangzData(LightningDataModule):
             train_batch_size: int = 32,
             eval_batch_size: int = 32,
             label_column: str = 'label',
-            encode_columns: list = ['text'],
+            encode_columns=None,
             **kwargs,
     ):
 
         super().__init__()
+        if encode_columns is None:
+            encode_columns = ['text']
         self.model_name_or_path = model_name_or_path
         self.task_name = task_name
         self.max_seq_length = max_seq_length
         self.train_batch_size = train_batch_size
         self.eval_batch_size = eval_batch_size
 
-        # self.text_fields = self.task_text_field_map[task_name]
-        # self.num_labels = self.glue_task_num_labels[task_name]
         self.tokenizer = AutoTokenizer.from_pretrained(self.model_name_or_path, use_fast=True)
         self.prepare_data_per_node = True
         self.n_cpu = 2
@@ -384,41 +394,48 @@ class TyqiangzData(LightningDataModule):
         self.encode_columns = encode_columns
 
         #  Tokenize the dataset
-        if not os.path.isfile(f'{self.task_name}_tokenized_data.pt'):
-            print("File not exist")
-            self.prepare_data()
+        self.prepare_data()
 
-    def clean_data(self, example):
-        example['sentence'] = ["{} {}".format(title, review) for title, review in
-                               zip(example['review_title'], example['review_body'])]
-        example['stars'] = [int(star) - 1 for star in example['stars']]
-        return example
-
+    # no need to clean data for this task so no self.clean() method
     def prepare_data(self):
         # called only on 1 GPU per node. Do not do any self.x assignments here
         # download, split tokenise data and save to disk
-        print(f'Download and Tokenise')
-        dataset = datasets.load_dataset(self.task_name, 'german')
 
-        for split in dataset.keys():
-            dataset[split] = dataset[split].map(self.encode_batch, batched=True)
-            remove_features = set(dataset[split].features) ^ set([self.label_column] + ["input_ids", "attention_mask"])
-            dataset[split] = dataset[split].remove_columns(remove_features)
-            dataset[split] = dataset[split].rename_column(self.label_column, "labels")
+        if not os.path.isfile(f'{self.task_metadata["tokenize_folder_name"]}/tokenized_data.pt'):
+            print("File not exist")
+            print(f'Download and Tokenise')
+            dataset = datasets.load_dataset(self.task_name, 'german')
+            for split in dataset.keys():
+                dataset[split] = dataset[split].map(self.encode_batch, batched=True)
+                remove_features = set(dataset[split].features) ^ set(
+                    [self.label_column] + ["input_ids", "attention_mask"])
+                dataset[split] = dataset[split].remove_columns(remove_features)
+                dataset[split] = dataset[split].rename_column(self.label_column, "labels")
 
-            # Transform to pytorch tensors and only output the required columns
-            # self.dataset[split].set_format(type="torch", columns=["input_ids", "attention_mask", "labels"])
+                # Transform to pytorch tensors and only output the required columns
+                columns = [c for c in dataset[split].column_names if c in self.loader_columns]
+                dataset[split].set_format(type="torch", columns=columns)
 
-            columns = [c for c in dataset[split].column_names if c in self.loader_columns]
-            dataset[split].set_format(type="torch", columns=columns)
+            # save the tokenized data to disk
+            try:
+                Path(f'{self.task_metadata["tokenize_folder_name"]}').mkdir(parents=True, exist_ok=True)
+                torch.save(dataset, f'{self.task_metadata["tokenize_folder_name"]}/tokenized_data.pt')
+            except:
+                print("File already exist")
 
-        # save the tokenized data to disk
-        torch.save(dataset, f'./{self.task_name}_tokenized_data.pt')
-        print(f'Dataset Tokenised')
+        else:
+            print("File exist. Load Tokenized data in setup.")
+
 
     def setup(self, stage: str):
         # load data here
-        self.dataset = torch.load('tokenized_data.pt')
+        try:
+            self.dataset = torch.load(f'{self.task_metadata["tokenize_folder_name"]}/tokenized_data.pt')
+        except:
+            print("File not exist")
+            self.prepare_data()
+            self.dataset = torch.load(f'{self.task_metadata["tokenize_folder_name"]}/tokenized_data.pt')
+
         self.eval_splits = [x for x in self.dataset.keys() if "validation" in x]
         print('dataset loaded')
 
@@ -444,23 +461,37 @@ class TyqiangzData(LightningDataModule):
                     self.eval_splits]
 
 
-if __name__ == "__main__":
-    dm = GlueModule("distilbert-base-uncased", task_name="sst2", max_seq_length=256, train_batch_size=32,
-                    eval_batch_size=32)
-    dm = AmazonMultiReview("distilbert-base-uncased",
-                           task_name="amazon_reviews_multi",
-                           max_seq_length=256,
-                           train_batch_size=32,
-                           label_column='stars',
-                           encode_columns=['review_body', 'review_title'])
-    dm.setup("fit")
+def getDataModule(task_name="", model_name_or_path: str = "distilbert-base-uncased",
+                  max_seq_length: int = 128, train_batch_size: int = 32,
+                  eval_batch_size: int = 32):
 
-    # dm = TyqiangzData("distilbert-base-uncased",
-    #                   task_name="tyqiangz/multilingual-sentiments",
-    #                   max_seq_length=256,
-    #                   train_batch_size=32,
-    #                   label_column='label',
-    #                   encode_columns=['text'])
-    # # dm.prepare_data()
-    # dm.setup("fit")
+    if task_name == "sst2":
+        return GLUEDataModule(model_name_or_path=model_name_or_path, task_name="sst2",
+                          max_seq_length=max_seq_length,
+                          train_batch_size=train_batch_size,
+                          eval_batch_size=eval_batch_size)
+
+    elif task_name == "amazon_reviews_multi":
+        return AmazonMultiReview(model_name_or_path=model_name_or_path,task_name="amazon_reviews_multi",
+                             max_seq_length=max_seq_length, train_batch_size=train_batch_size,
+                             eval_batch_size=eval_batch_size,
+                             label_column='stars',
+                             encode_columns=['review_body', 'review_title'])
+
+    elif task_name == "tyqiangz":
+        return TyqiangzData(model_name_or_path=model_name_or_path,
+                             task_name="tyqiangz/multilingual-sentiments",
+                             max_seq_length=max_seq_length,
+                             train_batch_size=train_batch_size,
+                             eval_batch_size=eval_batch_size,
+                             label_column='label',
+                             encode_columns=['text'])
+
+
+if __name__ == "__main__":
+    dm = getDataModule(task_name="tyqiangz",
+                       model_name_or_path="distilbert-base-uncased",
+                       max_seq_length=256, train_batch_size=32,eval_batch_size=32)
+    # dm.prepare_data()
+    dm.setup("fit")
     print(next(iter(dm.train_dataloader())))
